requirements.txt
google-genai>=0.1.0
pyautogui>=0.9.53
mss>=6.1.0
opencv-python>=4.5.0
numpy>=1.24.0
Pillow>=8.0.0
pyaudio>=0.2.11
websockets>=10.0
taskgroup>=0.1.0  # For Python < 3.11 async support
exceptiongroup>=1.0.0  # For Python < 3.11 async support
SpeechRecognition>=3.10.0
scipy>=1.11.0
sounddevice>=0.4.1
soundfile>=0.10.3.post1
python-dotenv>=0.19.0
pytesseract>=0.3.8
wave>=0.0.2
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.30.0
# System package - needs to be installed via system package manager
